refactor(ledger): collocate the v1 transaction route registrars :hammer:
The Huma terminal registrar came out of transaction_handler.go and the Fiber
guard-chain registrar came out of routes.go; both now live in
transaction_routes.go. The /v2 transaction surface is not a straight mirror of
/v1, so it keeps its own transaction_v2_register.go.

That empties routes.go of route surfaces: it now holds only the primitives the
per-resource files build their guard chains from, so its file comment says so
and the huma import goes.

No route, authz tuple, operation ID or path changed.
